No Runtime Panics
This change eliminates runtime panics in the runner and server. General
code hygine dictates that we should never crash for a recoverable
error. This commit removes use of `replicate/go/must` in the runner and
server.

Uses in main.go remain for the time being and will be removed in
future commits.

Additionally tests are out of scope for this change, and likewise should
have all use of the `must` module removed in favor of validation
utilizing assert/require.