Test harness retrofit: `coder_test.go`
Retrofit `coder_test.go`. This change required extra modifications as it
surfaced an issue with how PYTHONPATH was being constructed in the
`harness_test.go` helper file. Now it is possible, and recommended, to
pass the python binary through the handler instead of assuming PATH is
correct.

For compatbility and ensure normal PATH auto dectection works we also
fall back to the previous behavior of `"python3"` as a bare string.