Test Harness Retrofit: `path_test.go`
Retrofit `path_test.go`.

We no longer download from the public internet the test files. This
ensures our validation is clean and will not be impacted by upstream
changes.

Due to timeouts, it was highlightes that blocking on channels in tests
needed a timeout mechanism. This was added across all converted tests.