Test Harness Retrofit: `shutdown_test.go`
This is not a valid test scenario in unit tests. In the future a proper
integration test should cover this behavior. With the move to in-process
(idiomatic) testing, we do not want to interrupt the actual test
runners. Testing the shutdown path with http request to `shutdown` can
also be added in the future.

This commit removes `shutdown_test.go`