Linting: remediate `ineffassign`
Along with the minor changes, the handleInputPaths was both non-descript
and complext enough logic to warrant tests. Comprehensive tests are
added for the complex function. For clarity `handleInputPaths` ->
`proccessInputPaths` and we make it clear with the docstring that this
mutates in place. We maintian using the returned value for clarity.

Copying the input could be memory intensive and therefore we opt for the
less clear but safer approach.