bugfix: Ensure schema is loaded before processing
Schema wasn't loaded prior to processing the input paths this resulted
in a noop nil response. To ensure we always load the schema first we now
block on runner ready state before processing. Additionaly a nil doc is
considered an error. In the cases we call ProcessInputPaths we can
accept that error; in manager we explicitly check for doc being nil and
emit an error log.