Ensure that we always 202 for async
Async predictions should not 500. However, since we now known that the
runner failed before we send the prediction request down, we end up
needing to work around the error and still send a 202 *then* let the
webhook indicate failure.

While this behavior is somewhat crazy, it is the behavioral contract we
can revisit in the future.