bugfix: Ensure schema is loaded before processing (#221)
* bugfix: Ensure schema is loaded before processing

Schema wasn't loaded prior to processing the input paths this resulted
in a noop nil response. To ensure we always load the schema first we now
block on runner ready state before processing. Additionaly a nil doc is
considered an error. In the cases we call ProcessInputPaths we can
accept that error; in manager we explicitly check for doc being nil and
emit an error log.

* Ensure that we always 202 for async

Async predictions should not 500. However, since we now known that the
runner failed before we send the prediction request down, we end up
needing to work around the error and still send a 202 *then* let the
webhook indicate failure.

While this behavior is somewhat crazy, it is the behavioral contract we
can revisit in the future.